requirements.txt
--index-url https://pypi.python.org/simple/

### requirements.txt ##############################################
###
### To install, run:
###      $ pip install -r requirements.txt
###
###################################################################

requests
python-dateutil
-e .